Created Ball class framework and initializer
In this commit I created the framework we're going to start with for the ball class, with an initializer for the 4 major variables the ball needs: an x and y position, and an x and y velocity.

I also created placeholder methods for launch, which starts moving the ball at the start of a round, and bounce, which will run when the ball collides with another object.

Pygame implementation is not included yet, and will be done at a later time.