Completed a very basic version of the ball bounce
Added the functionality of bouncing the ball off of any block (paddle or walls). The current implementation is very basic and brute force, where when the ball hits a block, it just multiplies the ball's direction based speed by -1. 
A future implementation will be more efficient rather than looping through a range of coordinates (avoiding the brute force method).