Added the ball's functionality to be a polygon
Previously, the ball was made of a center point and a radius. This
results in us having only a circle as the only option for the ball.
With this commit, we use a center point and a list of points rather than the previous properties. This gives us the ability to add any polygon we want simply by making hardcoding them out of points.
This also gives the future possibility of a ball editor where in the menus, we can design our own ball. This is only if we have time to or as a call to action to our classmates.